Add hydep.internal.configmethod decorator
Decided to use ConfigParsers for the primary configuration
engine. Needed a way to allow versatile configuration [file-based,
and in-memory] library and chose to go with this one. Many solvers
will have a configure method that should be flexible enough
to handle these cases [config file or pass a dictionary].
This decorator decorates a method that anticipates the first
non-instance argument to be the configuration object, e.g.

class Demo:
        @configmethod
        def configure(self, cfg):
                # cfg is a configparser.ConfigParser
                pass

Accepts string or pathlib.Path as files to be read,
dict or Mappings, or ConfigParsers as well.