Write pins w/ burnable materials specially
The SerpentWriter now checks if a pin has any burnable materials
prior to writing the definition. If none are found, the "standard"
or simple way using the pin construct is used. Otherwise, we
need to build the pin such that the burnable materials occupy a
unique sub-universe so fission matrix, xs, and other physics can
be tallied.

Use a special routine for handling these specially, wherein burnable
materials are placed inside an infinite surface, and fill a special
cell. This special cell is an infinite universe of just the burnable
material, and occupies one ring of the pin placed in the pins
universe.  This way a command like "set gcu <fuelmat.id>" will tally
group constants in just this material. Same for fission matrix and
micro-depletion / local reaction rate settings.