Add registry for reproducible, unique class ids
It is often important to represent a specific material or geometric
item (pin cell, lattice, etc.) with a unique identifier. Most codes
require this to be a positive integer, e.g. universe id in a Monte
Carlo code. One could use ``hex(id(obj))`` to obtain a unique number,
but this is not reproducible, a desired attribute for testing.

To facilitate this, hydep.internal.registry provides three functions
``register``, ``unregister``, and ``get`` that all accept a single
argument representing the key (or more typically object type) to
be registered. These functions can and will be used to create
unique Material and Universe ids useful for writing Serpent, OpenMC,
and other transport solver inputs.

                               ** NOTE**

The registry has no knowledge of copies created using copy.copy
or copy.deepcopy. This can destroy the guarantee of uniqueness.