Apply new id registry to Material, Universe, and subclasses
Material and Universe objects now have an id created using
the registry system, helping to ensure uniqueness and
reproducibility. Changes had to be made to the various
recursive digging routines to not use object.id, as the ids
are only unique among Material and Universe types, not all
objects everywhere. Calls to id(object) replace the previous
object.id calls.

Instances of Material, Universes, and subclasses like Pin and
BurnableMaterial are registered using either Material or Universe
as the key, so that there will always be one Material (or subclass)
with an id of 1. To put it in another way, all materials found with
Universe.findMaterials() will have a unique id. This also helps
avoid clashes in geometry building with multiple universe N
running around.