Slighty better status check for SerpentRunner
The check=True argument to subprocess.run has been removed, because
if Serpent fails, then no process object is returned and there is
no standard out to be processed. Instead, the return code on the
process object is examined.

The output message is not parsed still, but it does print better
now that the values are decoded from binary to plain string