Update serpent pin write; pin universes as p<id>
Cleaned up / improved how the SerpentWriter handles writing
pins, burnable and non-burnable pins. Universes are now led with
a "p" before the universe id to make things a little more clear,
now that there are just a lot of plain numbers running around as
identifiers. Use material ids in place of names in a lot of places,
as the names can have spaces in them / potentially weird characters.
A minor comment is written before the pin declaration with the name,
if applicable.