Add a dedicated universe-based fission matrix parser
Can be imported as hydep.serpent.fmtx.parseFmtx, and
accepts a stream of readable data. The returned object has
two attributes: matrix and universes. Matrix is the fission
matrix, stored as a CSR matrix. Universes is the ordering
of universes such that burnable region i can be found at
universes.index(i)