Add SerpentProcessor for handling results, fmtx files
Capable of parsing result files for fluxes and macroscopic cross
sections in burnable regions, as well as multiplication factor.
Uses a small wrapper around serpentTools.read to do some filtering
of data: only group constant data, ABS_KEFF, and some GC metadata
(number of groups, group structure, etc.) is read from the file.
The processor can switch between collecting infinite medium or
B1 cross sections, but no configuration is currently provided.

A TransportResult is populated with the fluxes, [keff, unc], and
homogenized macroscopic cross sections from the result file.
This should be called first prior to processing other files,
which can then add their data onto this TransportResult.
The processor requires the universes in which all burnable materials
reside, and the ordering. This enables the fluxes and cross sections
to be correctly ordered.

The processor also handles the fission matrix file, primarily dumping
off to the parseFmtx function for the heavy lifting. A check is
performed to ensure that the ordering of the universes in the fission
matrix is consistent, but no corrective actions are taken.