Return outermost Cartesian Lattice identifier
Previously, if a Cartesian lattice was used in some other Universe,
and did have an outside, then the bounding universe would only
fill a location [position in larger Cartesian lattice or axial layer]
with the lattice definition, e.g. "lat <X>", not with the proper
universe definition, e.g. "
cell <X> <U> ...
cell <Y> <U> ...
"

This resulted in a warning from Serpent that a cell was not used,
but more crucially could result in non-defined regions. Consider
an assembly placed in a vertical lattice stack that is slightly
wider in the X and/or Y direction. This is where the outer material
should reside, but since the truly complete universe is not used,
the one created with cells, this region would remain un-filled.

This is resolved by changing what is returned from the Cartesian
lattice routine to return the "lat <X>" value only if the outer
material is not used. Otherwise, the lattice's ID is returned, since
that is the universe that contains both the lattice and its outer
material.