Provide SerpentSolver.hooks
Can be set directly, or with setHooks methods. Both perform
the same actions. Should only be set once, otherwise not all
the required input cards may be written, or data will tried
to be processed that does not exist.

Changes are forwarded along to associated helpers, currently
just the writer.