Teach SerpentSolver about SerpentProcessor
The transport result is now populated with fluxes and
multiplication factor from the primary result file. If the solver is
told to grab macroscopic cross sections, they are also pulled from
the output file. Inclusion of the fission matrix hook will populate
the TransportResult.fmtx attribute.

With this, it is now possible to create a full input file, run
Serpent, and process the results entirely using the hydep API
and also just with the SerpentSolver methods.