Implement SerpentWriter microxs hook on base file
Creates a block in the base file that creates a single
depletion step and writes all the mdep cards to all
burnable materials. The depletion step is a hack such that
the mdep cards actually do something. Otherwise these won't
work. An alternative would be to use a detector, but
that would require several dummy materials with a single isotope,
which we might also want to avoid. A tweak in the Serpent code
may be necessary to avoid the second redundant depletion step,
or to create a consolidated file with the cross sections
for depletion. That solution is to be determined.

The mdep cards are unique to each burnable material, and
contain all the isotopes in the depletion and transmutation chain
for a specific material. The search routine goes over all reactions
products and fission products, adding new ones to the stack.
This way we don't need to tally isotopes that won't show up in
the depletion chain, like uranium in a non-integral burnable
absorber like boroscillicate glass or a burnable coating