Avoid clashes when writing CartesianLattices
Previously, it was possible for the universe that contained a
Cartesian lattice to be written as a single number. The internal
"lat <>" card always had the modified name, but the resulting universe
could be written with a number. We want to reserve strictly numeric
universes for regions of burnable materials, so this had to be
resolved.

How, when a lattice is written, it is never placed in a purely numeric
universe.