Always use a single energy group in set nfg
For now, all homogenized data and microscopic cross sections will be
obtained using a single energy group [0, 20] MeV. Obviously this will
not work well with a diffusion code, but we will cross that bridge
later, as that entails some way to make a generic energy structure
that will play with all of Serpent's particularities, e.g. alignment
with underlying microgrid, various construction methods, etc.

The energy grid is simply called "energies" and is hard coded into
the base file. The reference test file is updated.