Always write hooks in main file - detector fluxes
Even if no homogenized macroscopic cross sections are requested,
the flux in each region is still necessary. Therefore a detector
setting is used to tally the flux in each burnable region. The tally
is created using the same group structure as the homogenized cross
sections, currently hard coded to a single energy group. This way
there should be a consistent energy structure, regardless of where
the information originates.

2x2 test file has been updated accordingly.