Provide SerpentProcessor.processDetectorFluxes
Pulls fluxes from detector file. Detector must be binned against
universe corresponding to burnable regions, and should be
consistent. There currently is no way to understand the universe
ordering in the detector file.

Detectors can also be binned against energy, but that is the
only other option. No checks are performed on the order of
the energy bins, as that should line up with the structure
of the other homogenized parameters, e.g. microscopic cross
sections.

The returned tallies are re-arranged such that the universe index
is first, as the TransportResult anticipates