Improve robustness, documentation on MicroXsVector
Fix a bug that could have allowed scaling a microxs vector
unintentionally when multiplying. The xs was passed directly
onto the new object, which was then scaled in place. Without
copying [now added], this would modify the original as well.
A test has been added to keep this bug from reappearing, hopefully.

Documentation on the main construction method and the preferred
class method MicroXsVector.fromLongFormVectors has been introduced.
Currently debating with myself if I really need this sparsity,
but it will be kept anyway.

The fromLongFormVectors has been cleaned up and improved by
allowing any Sequence of integers to come in, and then converted
on the way out. Copies are done, but this allows the sorting to take
place, and check to make sure all vectors are truly vectors of equal
size.