Use a helper method to validate Manager powers
Break out some of the logic from the init method onto a helper
that performs more robust checks on the power. If a single
non-sequence is given, it must be a positive real value. This will
be converted to a tuple of constant powers to all steps. Otherwise,
a sequence of positive reals must be passed, of equal length as the
number of time steps.