Provide ReducedOrderSolver.processBOS
Some solvers, like the SFV, may need information on the beginning
of step data. For example, the SFV must compute changes in macroscopic
cross sections from the beginnning of step and the unknown time point
for the prediction.

Subclasses are not expected to implement this, as it defaults to
a dummy method.