Make BurnableMaterial.index more robust - set once
Ensure that the index on a burnable material can only be set
once by making it a proper settable property. Type checks are
done to ensure the value is an integer (or can be coerced to one).
Negative values are not allowed, and no checks for uniqueness are
performed.

Two tests have been added to tests/test_material.py to pass some
additional values and data types to BurnableMaterial.index and
check for errors raised, correct setting, and that the value does
not change, even if the original variable is modified in place.