Update volumes with Model.differentiateBurnableMaterials
An additional argument is given to the method, denoting if the
user wants to update the volumes of burnable materials encountered.
This triggers Universe.countBurnableMaterials to compute the
number of times each material is encountered, and then scale the volume
accordingly. A check is performed to ensure the volume is provided
for each burnable material.