Assume Model is unbounded unless otherwise set
Previously, calling Model.bounds for a fresh model would
seek into the root universe and find the boundaries. Now, the model
boundaries must be set by the user, and will be considered unbounded
until the boundaries have been set. The documentation has been
updated, and users are now pointed towards Universe.boundaries() to
use the root universe to find boundaries. Notices are made that having
an unbounded Model could be problematic for some solvers and should be
avoided.