Introduce FISSION_YIELD feature
Indicate that a solver with this feature can produce [or a solver that
needs this capability requires] a single set of fission yields for
each isotope in each burnable region.

We can do some trickery and take all values from a single energy on
the depletion chain once solvers start implementing this.
Alternatively, and more desirably, follow some of the methods
included in openmc.deplete for better weighting of fission yields