Don't always write Serpent hooks
Partial revert of commit 4759c524, in that the SerpentWriter will
not always write hooks. The motivation for this is that, when actually
coupled, the writer will almost always have hooks {as I see the
project unfolding}. If there are any hooks given, then the flux
will still be obtained, with detectors or with local homogenization.

May fully revert this, or go further in the "always write something"
vein, but not yet.

The reference 2x2 lattice has been updated as well