Add test for testing configuration routines
Currently just tests the hydep.internal.configmethod
decorator, but may be populated later with other routines.

A basic class is created that has a test method decorated with
configmethod. The class also provides four data types that are
passed to the test method and compared to reference data. This
gives a good sense of the versatility of the decorator, handling
dictionaries, ConfigParser objects, string file names, and
Path objects.