Start SerpentSolver regression test - flux, keff
Create a regression test for the Serpent solver that utilizes a
2x2 lattice, but with different burnable materials in each region.
The solver is configured to use a dedicated random seed, currently
just 12345678910, to help ensure reproducible results.

The Solver goes through the main aspects of the interface:
1) beforeMain <- model and burnable materiasl
2) bosUpdate <- dummy concentrations, time 0, nominal power
3) execute
4) processResults -> TransportResult

The test relies on a temporary directory for the simulation, so the
base file and potential zips are not added to the user's working
directory. These may be helpful on failures, buuut that's for later.

Comparisons are performed using the ResultComparator so that
reference data is extracted in a "standardized" way. This also
allows the test writer to not worry about creating failure files
at the end.