Add serpent marker and apply to relevant tests
Specific tests can now be designated as part of a "serpent"
collection with the pytest.mark.serpent decorator.
Marking a test in such a way allows test runners to isolate / exclude
Serpent tests with
$ pytest -m serpent
$ pytest -m "not serpent"
respectively. This can be helpful down the road for removing tests
that a user does not need / can't run in the first place. Can't run
any Serpent tests if you don't have Serpent.