Allow ResultComparator to compare fission matrices
Works by converting to the sparse COO format, which can
be more easily written to / read from a long form file
using numpy.savetxt, as the data, row, and column vectors
will all be of equal size.

Comparisons are made in two approaches. First, if the row and column
index vectors are identical, then the arrays have the same structure
and location of non-zeros. The result of the comparison is then the
comparison of the sparse data vectors. Essentially [but not exactly],
if actual.row == ref.row and actual.col == ref.col:
        return actual.data == reference.data

Otherwise, and likely for larger problems, it is possible that
one matrix may have zeros where the other has very small values.
This may be due to the nature of [Monte Carlo] simulations, in that
some non-zero values are _statistically_ zero. Or it could be that
the arrays are not identical. So, if the underlying structures are
not identical, the dense arrays are compared. This hopefully
reduces the number of very large, dense (but truly sparse) matrices
running around.