Simply ResultComparator interface with main function
Accepts a TransportResult that will either be used to update
the reference data or test against the reference data. For the latter
case, an assertion is made that no tests have failed. This method
handles interpreting command line / pytest configuration options so
the test writer only has to create an instance of ResultComparator
and then run ResultComparator.main(res) with the test data.

This method has been applied to the Serpent 2x2 reference case.