Provide easier FissionYieldDistribution accessors
The following methods are provided to get around a potentially
expensive __getitem__ call that looks into the energy vector
and then returns the yields at that energy.

Provide ``at`` method that takes an index and pulls fission
yields for that energy index. Useful for taking lowest or highest
energies.

Provide ``get`` method that includes an absolute tolerance on the
requested key.

Provide better ``items`` and ``values`` generator methods that rely
on avoid __getitem__