Take constant fission yields in Serpent interface
Doesn't actually interact with Serpent output files at all.
Simply pulls the thermal yields from the isotopes. Can be configured
to pick any energy, and falls back to the thermal / lowest energy
yields if not found.

Returns a FakeSequence now to emulate a set of fission yields
for every burnable region.

Quick, and probably not too terrible for testing LWR problems. But,
since this is what I've working on for OpenMC in the past, I should
probably investigate this in the future.

Added to TODO file