Provide Manager.deplete
Depletes all burnable materials given starting point in time,
length of depletion interval, and previous transport result.
Uses multiprocessing to deplete using unique processes per
burnable material.

The final concentrations are ordered according to the
zaiOrder on the depletion chain for consistency. This means that
some materials will be depleted with a matrix that contains references
to isotopes that don't exist (U235 transmutation chain will appear
in non-mixed burnable absorbers). But these isotopes will [hopefully]
have zero concentration at the end, so the material object [or
whatever updates it] will be able to discern what to update.

Compared against Serpent depletion for a single pin cell and a
2x2 array of pins with two different fuels. Results are promising
and a regression test will be included shortly. The main issue in
the concentrations is likely the use of incorrect fission yields.

                                **TODO**

* Make the depletion solver configurable. Currently hard-coded
  to use Cram16
* Fission yield methodology needs to be updated