Overload Material.get: Look up correct isotope key
Calling Material.get(922350) previous would return None, even if
an isotope representing U235 existed on the material. This is because
keys are currently isotope objects, and an integer clearly doesn't
exist.

Now, a quick check and conversion is performed to ensure that the
key has some hope in existing before deferring back to dict.get