Add [finicky] depletion regression test
Uses a 2x2 toy model, with three partially burned fuel pins at
identical compositions and the fourth fresh. Materials and
models as made through test fixtures. A manager is produced,
and loaded with microscopic cross sections from text files.
Fluxes are also read in, while fission yields are taken
as thermal yields. The depletion step is taken to be 50 days.

The manager depletes all compositions, and then performs an update
of reference data, or makes a comparison to reference data. A
file-based comparison is probably not ideal, as many of the small
compositions (O(1E-20)) can have very small differences, yet this
causes the test to fail. I've found that running the regression test
alone will yield different results than in the full test suite.