Don't rely on attr: hydep.__version__ in setup.cfg
Very nice feature, but causes problems with fresh installs.
Essentially causes the package to be imported, which trickles
down to a numpy and/or scipy import. These are listed in the
install requires section, but the version search is done first.
Placing the explicit version in the setup.cfg means this has
to be updated at releases, but makes a completely fresh install
possible.

Related GH issues / workarounds:
https://github.com/pypa/setuptools/issues/1724
https://github.com/pypa/pip/issues/6350