Provide TimeTraveller for projecting data in time
The TimeTraveller represents values taken over time, using a deque
to support storing only X values at once. Time values are, with
as much energy as possible, ordered in a sorted manner. This is
controlled through append and extend methods enforcing ascending
values of time. The subclasses are responsible for checking
consistency/conformity of values. An insort method is provided
to inserting values and time points to maintain ordering.

The primary purpose of this class is to provide a __call__ method that
performs the projection, through extrapolation, interpolation, or some
other fitting. The modification methods are provided to assist with
searching and fitting. Using a deque is advantageous as it can be used
to shove out previously computed values, e.g. only use the three latest
points for extrapolation.

A CachedTimeTraveller is also provided that store the most recent value
to __call__ and computed quantity to avoid potentially expensive
calculations.