TemporalMicroXs now subclass of CachedTimeTraveller
Provides some more additional convenience methods, but the
bulk of the responsiblity for managing the insertion / appending
of time points is now left to the parent class.

The cache is used to store values of projected cross sections
at the same time point without having to perform multiple fittings.

The insort, append, and extend methods support passing either
MicroXsVector instances or numpy arrays (iterables for extend).
For vectors, the zai, zptr, and rxn data must be identical. It is
possible to work if the incoming data contains a subset of reactions,
but that has been left out. Arrays must simply have the same structure
as the underlying cross section matrices (number of reactions and number
of energy groups).