Provide NubarPolyFit utility in sfv module
SFV method needs a way to predict the new average number of
neutrons produced per fission. Current practice is to extrapolate
or interpolate based on previous time steps, where the nubar data
are homogenized in each burnable region. This is essentially like
a one-group "macroscopic" nubar number.

The NubarPolyFit class fulfills the TimeTraveller interface for
storing one-group nubar values in each burnable region. It is only
intended to be used by the SFV interface, and is tucked inside. Users
have control over number of previous points and fitting order,
defaulting to a total of three points, and linear fitting.

Unit tests have been added to explore the functionality of this class.