Make BeavrsMaterials a module-scoped fixture
I've noticed that some tests that rely on a fixture that provides
materials may fail if earlier tests fail. This primarily happens
for the writers, as a failed test does not destroy the materials,
causing the next materials to have higher ids than the tests expect.
This is a fault of the very rigid expected strings and/or the material
id system. This bug can be reproduced by adding an "assert False" line
into some of the Serpent writer tests.