[regression] Update Serpent 2x2 w/o batching
Fairly large change here, as two key items have changed in this
branch. First, the batching has been turned off due to how it
changes the INF_FLX value needed for SFV. Following this, commit
7458b19 changed the number of active cycles from 10 to 20 by doubling
the number of active batches for this regression test.

The fission matrix is mostly within statistics, while K and flux
are not. The change in flux is attributable to the changes in batching,
as described in this branch. For some reason, introducing batches into
the pop card reduces the value of INF_FLX.

The k change is more noticeable, but the uncertainties are still large
so the change is (almost laughably) statistically negligible.