Add note about removing flaky tests
Since the mere existence of material and universes can change
the tests, it is possible that failure in one test may mess up another.
A simple example is the Serpent writer tests, that look for strict
agreement in files and having a material with the wrong ID can screw
this up.