Configure substeps / divisions for Manager timesteps
Allow the user to provide either an integer or sequence of integers
that represent the divisions in each coarse step. Also provide
Manager.substeps as a sequence of integers expanded to the number
of depletion intervals.

Better explained as the number of transport simulations per depletion
interval, both high fidelity and reduced order. Or the number of
depletion solutions per interval. The number of reduced order
simulations per interval is one less than the number of divisions.

The number of substeps is also exposed as a read-only property.

A single integer has been used for the current manager test.