Add hydep.utils.compBundleFromMaterials helper
Function that can create a CompBundle given a sequence of
Materials (works currently with any Mapping[key, float]).
Useful for preparing the bundle for the BOL simulation,
and for testing. Supports passing a threshold that will not
write isotopes below a given threshold,

Currently, the Serpent regression tests error out because we
pass None rather than a CompBundle to the solver routine.
This function hopes to provide a consistent way to make the
bundle.

A unit tests has been added as well.