Parametrize xs extrapolation test
Ensure that the maximum allowed order will be used,
up to the order requested by the user. If two points are given,
constant and linear extrapolation can be done easily, but anything
higher is not optimal nor possible. Numpy allows this, but will raise
some warnings.

The test is parametrized against fitting order, and the expected
vector will be determined in the test given the fitting order.