Test off-centered interpolation with microxs
Previously interpolated directly to the mid-point of the
times, meaning that constant and linear interpolation would return
the same value. This made it difficult to determine if the fit was
truly using constant or linear interpolation.

The target time is now off-set from the center (1/4 of the way from
a to b) so that constant and linear interpolation should not return
the same value. The expected microscopic cross sections are determined
by the order provided to the test.