Implement the coupled sequence in Problem.solve
Relies on two helper methods to walk through all depletion
steps and substeps for a given coarse step.

No testing is included at the moment, but this has been run with
good success on the cluster. The results needs to be investigated.
But, the solvers all talk to each other and pass data around.

This is the big one friends.