Use logging module for updates during Problem.solve
Can be retrieved and configured with logging.getLogger("hydep").
Prints info messages before and after each transport solution,
given the Solver class name, time step, substep index (if applicable),
time in days, and multiplication factor.

If the multiplication factor computed by the reduced order solver
is all nans, for the case of the SimpleROSolver, these values
are not printed.