Pass initial CompBundle to Manager.deplete
Used to instruct the depletion chain where to put isotopes,
and don't rely on updating material compositions in memory. This way,
the output of one depletion run can be passed directly into the next
high-fidelity solver, and back into the next depletion step.

The regression test has been updated to use compBundleFromMaterials
for make the initial concentrations and passed to Manager.deplete