Provide and test serpent.utils.findProblemIsotopes
Responsible for digging through the xsdata file and looking for
isotopes that exist under different (Z, A) identifiers,
(Am242_m1 -> 95342), and isotopes that don't exist in the library.
This is useful because currently, metastable isotopes are written
only using their ZA number, causing a double count as both
Am242 stable and Am242 would be written in the material block
under 95242. This is currently allowed in Serpent
https://ttuki.vtt.fi/serpent/viewtopic.php?f=25&t=3323