Provide SerpentWriter.writeMatIsoDef
Writes just the isotopic block for a material in both
the base file and each subsequent steady state files. Uses
the problematic isotope record to determine which isotopes
should do not exist and which should be replaced. Metastable
Am242 with ZAI 952421 goes to written ZA 95342.

Public so it can be reliably tested, but most of these methods
should never be directly called by the user.