Introduce FissionYieldFetcher for getting yields from Serpent
Inspired by OpenMC AveragedFissionYieldHelper, in that it computes the
fission-rate-weighted fission yields for isotopes with multiple sets
of yields. The fission rate is computed in bins surrounding fission
yields, using a mid-point lethargy scheme. The fission yields are
then weighted by the ratio of (sigf_g * phi_g) / sum (sigf_i * phi_i)

This class provides inputs that can be written to the main input file
that create the detector commands. Similarly, it is responsible for
processing those detectors, and weighting the yields.