[regression] [major] Update depletion with better chain
Large update here, in particular for isotopes that are produced /
destroyed primarily through decay reactions. Pu-240 has increased
in the "fresh" pin, because previous commits correctly handle these
decay chains.

Some value are negative, on the order of -1E-40 or smaller. A todo
item has been added to handle this. The Serpent writer already skips
isotopes with density less than zero, so we're good there. But it could
be problematic down the road.