Only open HdfStore when writing; other improvements
This helps reduce the risk that early termination will corrupt
the result files by only opening the file in the few writing methods.

Also flattened out the code by removing some of the internal methods
used for writing / prepping data sets. These aren't too cumbersome nor
lengthy and have been moved into the __init__, beforeMain,
postTransport, or writeCompositions methods

The HdfStore also checks if the result file exists prior to writing.
If the file does exist and is not a file, then an error is raised.
Overwriting a file is controlled by existOkay parameter, defaulting to
True. This causes an error or warning. Testing is added for these
cases