Provide indexing information to Stores at beforeMain
User's don't have to know all the information at the beginning
of the problem, as differentiating burnable materials may produce
a difficult to determine amount. This way, the framework will be
responsible for producing allocation information, like number of
transport steps, (coarse and total), groups, and number of burnable
materials.