Require TimeTraveller to provide popleft
Turns out deque.index will raise an IndexError if the bounded deque
would exceed the limits. Some logic has been added to try and avoid this
case. First, if the item would be placed at the end, use append. If the
deque is unbounded or not at the max length, then insert away.

If the item would be placed at the end of a full, bounded deque,
and subsequently dropped off, then it is not inserted. Finally, if the
item would go somewhere in the middle, popleft is called so the lowest
value is removed, and the item is inserted one position less.