Provide HydepSettings.updateAll
Expects a mapping of settings, with at least one level that
is ``hydep`` to be passed to HydepSettings.update. Other levels use
the dynamic subsection lookup to match settings like hydep.<key> to the
corresponding SubSetting. The documentation has been updated, and the
test has been modified to examine this feature.