Provide [bos|eol]Solve for SerpentSolver, CoupledSerpentSolver
Lost the archiving capabilities, but that has become more complicated
with the rundir / basedir / tempdir settings. Will fold that back in
soon.

The methods are pretty simple, and less information is passed between
methods now / stored on the instances. Will want to wrap the solution
steps inside Problem with try / excepts for termination and archiving
eventually.