Provide various Enumerations for accessing Hdf data
Three enumerations are provided that describe the stored
data. They are:

HdfStrings: names of primary datasets and groups (useful for iteration)
HdfSubStrings: names of sub-groups
HdfAttrs: Attributes for the base attrs manager

It is the intent that these strings will be more backwards compatible
and flexible than having to track down strings constantly.