Start documenting the API
Basically a bunch of summary tables describing the API. There are some
errors in docstrings that this is revealing, as well as some typos.
These will be adjusted in the future.

Nothing is included regarding examples nor methodology. The
configuration is extremely simple, little edits to what is produced
from sphinx-quickstart