Provide and test Model.isBounded
Supports checking all dimensions, or just a single dimension. Returns
a boolean indicating if the model is sufficiently bounded in one
or all dimensions.

When checking all dimensions, the z dimension is neglected. This
is done to allow support for 2D models that do not need to be bounded
in the Z direction.