Pull bounds from Model or Model.root, or raise GeometryError
The SerpentWriter will now first attempt to pull the model boundaries
from the model, then model.root.

The internal mechanisms for writing cell boundaries have been
re-worked to support this flexibility. Removed the TODO item on
pulling bounds from model, not model.root

Added a unit test that checks 1) GeometryError is raised if the
geometry is unbounded and 2) checks that identical files are written
written if model.bounds is provided and mode.root.bounds is provided but
model.bounds is not