Use sfv 0.3.2: Direct access to FORTRAN SFV routine
New release provides sfv.lib with a very small wrapper around
the predict_spatial_flux function. Provided arrays are of the right
datatype, this routine gets around all the error checking, conversions,
and type checks. It is more dangerous, yes. But the SfvSolver
takes care to ensure the vectors are all of the same shape and
data type by placing them in large array anyway.