Require Manager when writing Serpent inputs -> reaction index
Use DepletionChain.getReactionIndex to get the zai, mt numbers
for the mdep commands. This requires the manager to be passed
to more of the writing methods, and specifically the chain.
Changed methods include:

- SerpentWriter.writeBaseFile
- ExtDepWriter.writeCouplingFile
- BaseWriter.writeMainFile

This does remove the need to search through all isotopes and
reactions across __all__ materials, which is a pain. Instead,
the text string can be made once and applied across all
materials.

The upper limit ZAI=992550 is maintained.