Add hydep.sfv.MixedRK4Integrator
Previous experience has demonstrated that the SFV method
does not perform well for very small steps. This is likely related
to the small change in macroscopic absorption cross sections ruining
the prediction. When using the RK4 scheme, two solutions are obtained
by depleting to the midpoint. If the initial stages of a problem are
modeled using single day steps, this means there will be half-day
solutions and possibly a small enough change in cross sections to
ruin the prediction.

This integrator allows the user to use some other scheme, e.g.
predictor, for the small steps and the RK4 for larger steps.
The default cut off is five days, below which the predictor
scheme will be used. Otherwise the RK4 scheme will march across
the depletion interval.