Boundary extents are now accessible by attribute
Each dimension on a Boundaries class has a lower and upper
attribute that can be accessed with bounds.<d>.lower and
bounds.<d>.upper. This is enabled with a secondary class,
Dimension, that just handles the lower and upper boundary
extents. Acts like a tuple but with some more sugar for value
checking (making sure lower < upper, reals, etc.)