Add jupyter notebook-based examples
Model a 2D pin cell and then perform some post-processing. On my
current laptop, the final Serpent simulation fails due to a memory
allocation error. This means that there is a failure inside the
pincell model, which doesn't look great. But also the result file
generated is incomplete.

This is why there are some strange time-slicing calls to avoid the
last time step which has some erroneous data