Improve error parsing of Serpent error
A list of patterns is used to parse the Serpent error stream
and determine the source of error. Currently the only pattern is
an out of memory error, but the list can easily be appended later.

A helper method on the base runner is used by the SerpentSolver to
process the stdout stream, and raise the failed solver error